requirements.txt
MetaTrader5
pandas>=2.2.0
numpy>=1.26.4
requests>=2.31.0
fredapi>=0.5.1
yfinance>=0.2.33
scikit-learn>=1.4.0
xgboost>=2.0.3
lightgbm>=4.1.0
ta>=0.11.0
python-dotenv>=1.0.0
schedule>=1.2.1
loguru>=0.7.2
pyarrow>=14.0.2
matplotlib>=3.8.2
seaborn>=0.13.0
joblib>=1.3.2